pyproject.toml
[tool.poetry]
name = "work-shift-service"
description = "Work Shift Service"
version = "0.1.0"
readme = "README.md"
authors = ["myapp.com"]


[tool.poetry.dependencies]
python = "3.9.13"

django = { version = "^3.2", extras = ["argon2"] }
django-split-settings = "^1.1"
django-axes = "^5.35"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.12"
django-stubs-ext = "^0.5"
django-extensions = "^3.1.5"

psycopg2-binary = "^2.9"
gunicorn = "^20.0"
python-decouple = "^3.6"
structlog = "^21.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.22.1"
djangorestframework-stubs = "^1.7.0"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.4"
django-querycount = "^0.7"
django-migration-linter = "^4.1"
django-extra-checks = "^0.11"
django-coverage-plugin = "^2.0"
nplusone = "^1.0"

wemake-python-styleguide = "^0.16"
flake8-pytest-style = "^1.6"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
nitpick = "^0.32"

pytest = "^7.1"
pytest-django = "^4.5"
pytest-cov = "^3.0"
pytest-randomly = "^3.12"
pytest-deadfixtures = "^2.2"
pytest-timeout = "^2.1"
django-test-migrations = "^1.2"
hypothesis = "^6.47"

mypy = "^0.961"
django-stubs = "^1.12"

sphinx = "^5.0"
sphinx-autodoc-typehints = "^1.18"
tomlkit = "^0.11"
doc8 = "^0.11"

yamllint = "^1.26"
safety = "^1.10"
dotenv-linter = "^0.3"
polint = "^0.4"
dennis = "^1.0"
dump-env = "^1.3"
ipython = "^8.4"
factory-boy = "^3.2.1"
responses = "^0.21.0"
mock = "^4.0.3"
freezegun = "^1.2.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"
